Refactor demo app removing AbstractAttachmentQueue and related code in favour of using the @journeyapps/powersync-attachments lib
